made namespace bt, moved classes to bt::, move all classes from main file to their own
