+++ battle_tanks/client/headers/game/entity/game_entities.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "game/entity/game_object_type.hpp"
+#include "game/entity/game_object.hpp"
+#include "game/entity/bullet.hpp"
+#include "game/entity/game_scene.hpp"
