+++ battle_tanks/client/headers/game/entity/game_object_type.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace bt
+{
+    enum class game_object_type : unsigned
+    {
+        unknown = 0,
+        tank,
+        bullet,
+        wall,
+        brick
+    };
+
+}
