+++ battle_tanks/client/headers/physics/game_object_b2d_link.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "game/entity/game_object_type.hpp"
+
+namespace bt
+{
+    struct game_object_b2d_link
+    {
+        game_object_type type{ game_object_type::unknown };
+        bt::uuid id{ 0 };
+    };
+}
