+++ battle_tanks/client/headers/renderer/container.hpp
@@ -8,20 +8,23 @@
 
 #include "render_object.hpp"
 
-class container : public render_object
+namespace bt
 {
-public:
-    virtual ~container() override = default;
+    class container : public render_object
+    {
+    public:
+        virtual ~container() override = default;
 
-public:
-    void add_child(const std::shared_ptr<render_object>& child);
+    public:
+        void add_child(const std::shared_ptr<render_object>& child);
 
-    void remove_child(const std::shared_ptr<render_object>& child);
-    void remove_child(const render_object* const child);
+        void remove_child(const std::shared_ptr<render_object>& child);
+        void remove_child(const render_object* const child);
 
-protected:
-    void draw(const std::shared_ptr<sf::RenderTarget>& render_target, const sf::Transform& parent_transform) const override;
+    protected:
+        void draw(const std::shared_ptr<sf::RenderTarget>& render_target, const sf::Transform& parent_transform) const override;
 
-protected:
-    std::vector<std::shared_ptr<render_object>> children_{};
-};
+    protected:
+        std::vector<std::shared_ptr<render_object>> children_{};
+    };
+}
