+++ battle_tanks/client/headers/renderer/renderer.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "renderer/textures.hpp"
+
+#include "renderer/render_object.hpp"
+#include "renderer/container.hpp"
+#include "renderer/sprite.hpp"
+#include "renderer/render_scene.hpp"
