+++ battle_tanks/client/headers/utils/uuid.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace bt
+{
+    typedef long long unsigned uuid;
+
+    uuid generate_uuid();
+
+}
