+++ battle_tanks/client/sources/renderer/container.cpp
@@ -1,7 +1,7 @@
 #include "renderer/container.hpp"
 #include "renderer/render_object.hpp"
 
-void container::add_child(const std::shared_ptr<render_object>& child)
+void bt::container::add_child(const std::shared_ptr<render_object>& child)
 {
     if (child->parent_ != nullptr)
     {
@@ -12,7 +12,7 @@ void container::add_child(const std::shared_ptr<render_object>& child)
     child->parent_ = this;
 }
 
-void container::remove_child(const std::shared_ptr<render_object>& child)
+void bt::container::remove_child(const std::shared_ptr<render_object>& child)
 {
     for (auto it = children_.begin(); it != children_.end(); ++it)
     {
@@ -24,7 +24,7 @@ void container::remove_child(const std::shared_ptr<render_object>& child)
     }
 }
 
-void container::remove_child(const render_object* const child)
+void bt::container::remove_child(const render_object* const child)
 {
     for (auto it = children_.begin(); it != children_.end(); ++it)
     {
@@ -36,7 +36,7 @@ void container::remove_child(const render_object* const child)
     }
 }
 
-void container::draw(const std::shared_ptr<sf::RenderTarget>& render_target, const sf::Transform& parent_transform) const
+void bt::container::draw(const std::shared_ptr<sf::RenderTarget>& render_target, const sf::Transform& parent_transform) const
 {
     const sf::Transform transform = parent_transform * getTransform();
     for (const auto& ch : children_)
