+++ battle_tanks/client/sources/renderer/render_object.cpp
@@ -2,7 +2,7 @@
 #include "renderer/container.hpp"
 
 
-void render_object::free()
+void bt::render_object::free()
 {
     if (parent_ != nullptr)
     {
