+++ battle_tanks/client/sources/util/uuid.cpp
@@ -1,0 +1,12 @@
+#include "utils/uuid.hpp"
+
+namespace
+{
+    bt::uuid current_uuid{ 1 };
+}
+
+
+bt::uuid bt::generate_uuid()
+{
+    return current_uuid++;
+}
